news, tournaments, web api, DAL, footer,

+++ uksl.DAL/Entities/Person.cs
@@ -9,13 +9,13 @@ namespace uksl.DAL.Entities
     public class Person
     {
         public string AspNetUserId { get; set; }
-        public int Id { get; set; }
+        public decimal Id { get; set; }
         public string FName { get; set; }
         public string MName { get; set; }
         public string LName { get; set; }
         public DateTime? BirthDate { get; set; }
         public string NickName { get; set; }
-        public int? UniversityId { get; set; }
+        public decimal? UniversityId { get; set; }
         public string PhoneNumber { get; set; }
         public string Email { get; set; }
         public string TeamName { get; set; }
